extract EditorContextProvider
