+++ spx-gui/src/components/editor/panels/todo/AssetAddBtn.vue
@@ -141,7 +141,7 @@ import { addAsset, IsPublic, AssetType } from '@/apis/asset'
 import { useI18n } from 'vue-i18n'
 import { isImage, isSound } from '@/utils/utils'
 import { useNetwork } from '@/utils/network'
-import { useEditorCtx } from '@/components/editor/ProjectEditor.vue'
+import { useEditorCtx } from '@/components/editor/EditorContextProvider.vue'
 import { stripExt } from '@/utils/path'
 import { Backdrop } from '@/models/backdrop'
 import { Costume } from '@/models/costume'
